Add img/jambalaya.jpg, edit jambalaya.html
Created the img directory within odin-recipes. Added jambalaya.jpg to use within jambalaya.html. Edited jambalaya.html to populate the page with an image of the completed didsh, a description of the dish, an ingredients list, a step-by-step guide to cooking the recipe, and a link back to the recipe list.